# docs/openapi.json
{
   "components" : {
      "schemas" : {
...
         "DisabledEndpointError" : {
            "additionalProperties" : false,
            "properties" : {
               "error" : {
                  "enum" : [
                     "disabled-endpoint"
                  ]
               },
               "message" : {
                  "enum" : [
                     "This endpoint is disabled"
                  ]
               },
               "status" : {
                  "enum" : [
                     404
                  ]
               }
            },
            "type" : "object"
         },
...
         "DisabledUserError" : {
            "additionalProperties" : false,
            "properties" : {
               "error" : {
                  "enum" : [
                     "disabled-user"
                  ]
               },
               "message" : {
                  "enum" : [
                     "User is disabled"
                  ]
               },
               "status" : {
                  "enum" : [
                     401
                  ]
               }
            },
            "type" : "object"
         },
...
         "InvalidRequestError" : {
            "additionalProperties" : false,
            "properties" : {
               "error" : {
                  "enum" : [
                     "invalid-request"
                  ]
               },
               "message" : {
                  "enum" : [
                     "The request payload is incorrect"
                  ]
               },
               "status" : {
                  "enum" : [
                     400
                  ]
               }
            },
            "type" : "object"
         },
...
         "Mfa" : {
            "additionalProperties" : false,
            "properties" : {
               "ticket" : {
                  "type" : "string"
               }
            },
            "type" : "object"
         },
...
         "SessionPayload" : {
            "additionalProperties" : false,
            "properties" : {
               "mfa" : {
                  "$ref" : "#/components/schemas/Mfa"
               },
               "session" : {
                  "$ref" : "#/components/schemas/Session"
               }
            },
            "type" : "object"
         },
         "SignInTelegramSchema" : {
            "additionalProperties" : false,
            "properties" : {
               "initData" : {
                  "example" : "user%3D%257B%2522id%2522%32....85669%252C%2522first_name%2522%253A%2522%25D0%...",
                  "type" : "string"
               },
               "locale" : {
                  "default" : "en",
                  "description" : "A two-characters locale",
                  "example" : "en",
                  "maxLength" : 2,
                  "minLength" : 2,
                  "type" : "string"
               },
               "botId" : {
                  "example" : "1234567890",
                  "type" : "string"
               }
            },
            "required" : [
               "initData",
               "botId"
            ],
            "type" : "object"
         },
         "SignInAnonymousSchema" : {
            "additionalProperties" : false,
            "properties" : {
               "displayName" : {
                  "example" : "John Smith",
                  "type" : "string"
               },
               "locale" : {
                  "default" : "en",
                  "description" : "A two-characters locale",
                  "example" : "en",
                  "maxLength" : 2,
                  "minLength" : 2,
                  "type" : "string"
               },
               "metadata" : {
                  "additionalProperties" : false,
                  "default" : {},
                  "example" : {
                     "firstName" : "John",
                     "lastName" : "Smith"
                  },
                  "properties" : {},
                  "type" : "object"
               }
            },
            "type" : "object"
         },
         "SignInEmailPasswordSchema" : {
            "additionalProperties" : false,
            "properties" : {
               "email" : {
                  "description" : "A valid email",
                  "example" : "john.smith@nhost.io",
                  "format" : "email",
                  "type" : "string"
               },
               "password" : {
                  "example" : "Str0ngPassw#ord-94|%",
                  "type" : "string"
               }
            },
            "required" : [
               "email",
               "password"
            ],
            "type" : "object"
         },
         "SignInMfaTotpSchema" : {
            "additionalProperties" : false,
            "properties" : {
               "otp" : {
                  "type" : "string"
               },
               "ticket" : {
                  "example" : "mfaTotp:e08204c7-40af-4434-a7ed-31c6aa37a390",
                  "pattern" : "mfaTotp:\\b[0-9a-f]{8}\\b-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-\\b[0-9a-f]{12}\\b",
                  "type" : "string"
               }
            },
            "required" : [
               "ticket",
               "otp"
            ],
            "type" : "object"
         },
         "SignInOtpSchema" : {
            "additionalProperties" : false,
            "properties" : {
               "otp" : {
                  "type" : "string"
               },
               "phoneNumber" : {
                  "type" : "string"
               }
            },
            "required" : [
               "phoneNumber",
               "otp"
            ],
            "type" : "object"
         },
         "SignInPasswordlessEmailSchema" : {
            "additionalProperties" : false,
            "properties" : {
               "email" : {
                  "description" : "A valid email",
                  "example" : "john.smith@nhost.io",
                  "format" : "email",
                  "type" : "string"
               },
               "options" : {
                  "additionalProperties" : false,
                  "properties" : {
                     "allowedRoles" : {
                        "default" : [
                           "me",
                           "user"
                        ],
                        "example" : [
                           "me",
                           "user"
                        ],
                        "items" : {
                           "oneOf" : [
                              {
                                 "enum" : [
                                    "me"
                                 ]
                              },
                              {
                                 "enum" : [
                                    "user"
                                 ]
                              }
                           ]
                        },
                        "type" : "array"
                     },
                     "defaultRole" : {
                        "default" : "user",
                        "enum" : [
                           "me",
                           "user"
                        ],
                        "example" : "user",
                        "type" : "string"
                     },
                     "displayName" : {
                        "example" : "John Smith",
                        "type" : "string"
                     },
                     "locale" : {
                        "default" : "en",
                        "description" : "A two-characters locale",
                        "example" : "en",
                        "maxLength" : 2,
                        "minLength" : 2,
                        "type" : "string"
                     },
                     "metadata" : {
                        "additionalProperties" : false,
                        "default" : {},
                        "example" : {
                           "firstName" : "John",
                           "lastName" : "Smith"
                        },
                        "properties" : {},
                        "type" : "object"
                     },
                     "redirectTo" : {
                        "default" : "https://my-app.com",
                        "example" : "https://my-app.com/catch-redirection",
                        "type" : "string"
                     }
                  },
                  "type" : "object"
               }
            },
            "required" : [
               "email"
            ],
            "type" : "object"
         },
...
         "SignInVerifyWebauthnSchema" : {
            "additionalProperties" : false,
            "properties" : {
               "credential" : {
                  "additionalProperties" : false,
                  "properties" : {},
                  "type" : "object"
               },
               "email" : {
                  "description" : "A valid email",
                  "example" : "john.smith@nhost.io",
                  "format" : "email",
                  "type" : "string"
               }
            },
            "required" : [
               "email",
               "credential"
            ],
            "type" : "object"
         },
         "SignInWebauthnSchema" : {
            "additionalProperties" : false,
            "properties" : {
               "email" : {
                  "description" : "A valid email",
                  "example" : "john.smith@nhost.io",
                  "format" : "email",
                  "type" : "string"
               }
            },
            "required" : [
               "email"
            ],
            "type" : "object"
         },
...
         "UnauthorizedError" : {
            "additionalProperties" : false,
            "properties" : {
               "error" : {
                  "type" : "string"
               },
               "message" : {
                  "type" : "string"
               },
               "status" : {
                  "enum" : [
                     401
                  ]
               }
            },
            "required" : [
               "error",
               "status"
            ],
            "type" : "object"
         },
...
      },
      "securitySchemes" : {
         "BearerAuth" : {
            "scheme" : "bearer",
            "type" : "http"
         }
      }
   },
   "info" : {
      "description" : "Authentication for Hasura",
      "license" : {
         "name" : "MIT",
         "url" : ""
      },
      "termsOfService" : "",
      "title" : "Hasura auth",
      "version" : "0.17.1"
   },
   "openapi" : "3.0.0",
   "paths" : {
...
      "/signin/anonymous" : {
         "post" : {
            "deprecated" : false,
            "parameters" : [],
            "requestBody" : {
               "content" : {
                  "application/json" : {
                     "schema" : {
                        "$ref" : "#/components/schemas/SignInAnonymousSchema"
                     }
                  }
               },
               "description" : "",
               "required" : true
            },
            "responses" : {
               "200" : {
                  "content" : {
                     "application/json" : {
                        "schema" : {
                           "$ref" : "#/components/schemas/SessionPayload"
                        }
                     }
                  },
                  "description" : "User successfully authenticated"
               },
               "400" : {
                  "content" : {
                     "application/json" : {
                        "schema" : {
                           "$ref" : "#/components/schemas/InvalidRequestError"
                        }
                     }
                  },
                  "description" : "The payload is invalid"
               },
               "404" : {
                  "content" : {
                     "application/json" : {
                        "schema" : {
                           "$ref" : "#/components/schemas/DisabledEndpointError"
                        }
                     }
                  },
                  "description" : "The feature is not activated"
               }
            },
            "security" : [],
            "summary" : "Anonymous authentication",
            "tags" : [
               "Authentication"
            ]
         }
      },
      "/signin/telegram" : {
         "post" : {
            "deprecated" : false,
            "parameters" : [],
            "requestBody" : {
               "content" : {
                  "application/json" : {
                     "schema" : {
                        "$ref" : "#/components/schemas/SignInTelegramSchema"
                     }
                  }
               },
               "description" : "",
               "required" : true
            },
            "responses" : {
               "200" : {
                  "content" : {
                     "application/json" : {
                        "schema" : {
                           "$ref" : "#/components/schemas/SessionPayload"
                        }
                     }
                  },
                  "description" : "User successfully authenticated"
               },
               "400" : {
                  "content" : {
                     "application/json" : {
                        "schema" : {
                           "$ref" : "#/components/schemas/InvalidRequestError"
                        }
                     }
                  },
                  "description" : "The payload is invalid"
               },
               "404" : {
                  "content" : {
                     "application/json" : {
                        "schema" : {
                           "$ref" : "#/components/schemas/DisabledEndpointError"
                        }
                     }
                  },
                  "description" : "The feature is not activated"
               }
            },
            "security" : [],
            "summary" : "Telegram authentication",
            "tags" : [
               "Authentication"
            ]
         }
      },
      "/signin/email-password" : {
         "post" : {
            "deprecated" : false,
            "parameters" : [],
            "requestBody" : {
               "content" : {
                  "application/json" : {
                     "schema" : {
                        "$ref" : "#/components/schemas/SignInEmailPasswordSchema"
                     }
                  }
               },
               "description" : "",
               "required" : true
            },
            "responses" : {
               "200" : {
                  "content" : {
                     "application/json" : {
                        "schema" : {
                           "$ref" : "#/components/schemas/SessionPayload"
                        }
                     }
                  },
                  "description" : "Signed in successfully"
               },
               "400" : {
                  "content" : {
                     "application/json" : {
                        "schema" : {
                           "$ref" : "#/components/schemas/InvalidRequestError"
                        }
                     }
                  },
                  "description" : "The payload is invalid"
               },
               "401" : {
                  "content" : {
                     "application/json" : {
                        "schema" : {
                           "$ref" : "#/components/schemas/UnauthorizedError"
                        }
                     }
                  },
                  "description" : "Invalid email or password, or user is not verified"
               },
               "404" : {
                  "content" : {
                     "application/json" : {
                        "schema" : {
                           "$ref" : "#/components/schemas/DisabledEndpointError"
                        }
                     }
                  },
                  "description" : "The feature is not activated"
               }
            },
            "security" : [],
            "summary" : "Authenticate with email + password",
            "tags" : [
               "Authentication"
            ]
         }
      },
      "/signin/mfa/totp" : {
         "post" : {
            "deprecated" : false,
            "parameters" : [],
            "requestBody" : {
               "content" : {
                  "application/json" : {
                     "schema" : {
                        "$ref" : "#/components/schemas/SignInMfaTotpSchema"
                     }
                  }
               },
               "description" : "",
               "required" : true
            },
            "responses" : {
               "200" : {
                  "content" : {
                     "application/json" : {
                        "schema" : {
                           "$ref" : "#/components/schemas/SessionPayload"
                        }
                     }
                  },
                  "description" : "User successfully authenticated"
               },
               "400" : {
                  "content" : {
                     "application/json" : {
                        "schema" : {
                           "$ref" : "#/components/schemas/InvalidRequestError"
                        }
                     }
                  },
                  "description" : "The payload is invalid"
               },
               "404" : {
                  "content" : {
                     "application/json" : {
                        "schema" : {
                           "$ref" : "#/components/schemas/DisabledEndpointError"
                        }
                     }
                  },
                  "description" : "The feature is not activated"
               }
            },
            "security" : [],
            "summary" : "Sign in with a Time-base One-Time Password (TOTP) ticket",
            "tags" : [
               "Authentication"
            ]
         }
      },
      "/signin/passwordless/email" : {
         "post" : {
            "deprecated" : false,
            "parameters" : [],
            "requestBody" : {
               "content" : {
                  "application/json" : {
                     "schema" : {
                        "$ref" : "#/components/schemas/SignInPasswordlessEmailSchema"
                     }
                  }
               },
               "description" : "",
               "required" : true
            },
            "responses" : {
               "200" : {
                  "content" : {
                     "application/json" : {
                        "schema" : {
                           "type" : "string"
                        }
                     }
                  },
                  "description" : "Email sent successfully"
               },
               "400" : {
                  "content" : {
                     "application/json" : {
                        "schema" : {
                           "$ref" : "#/components/schemas/InvalidRequestError"
                        }
                     }
                  },
                  "description" : "The payload is invalid"
               },
               "401" : {
                  "content" : {
                     "application/json" : {
                        "schema" : {
                           "$ref" : "#/components/schemas/DisabledUserError"
                        }
                     }
                  },
                  "description" : "User is disabled"
               },
               "404" : {
                  "content" : {
                     "application/json" : {
                        "schema" : {
                           "$ref" : "#/components/schemas/DisabledEndpointError"
                        }
                     }
                  },
                  "description" : "The feature is not activated"
               }
            },
            "security" : [],
            "summary" : "Email passwordless authentication",
            "tags" : [
               "Authentication"
            ]
         }
      },
      "/signin/passwordless/sms" : {
         "post" : {
            "deprecated" : false,
            "parameters" : [],
            "requestBody" : {
               "content" : {
                  "application/json" : {
                     "schema" : {
                        "$ref" : "#/components/schemas/SignInPasswordlessSmsSchema"
                     }
                  }
               },
               "description" : "",
               "required" : true
            },
            "responses" : {
               "200" : {
                  "content" : {
                     "application/json" : {
                        "schema" : {
                           "type" : "string"
                        }
                     }
                  },
                  "description" : "SMS sent successfully"
               },
               "400" : {
                  "content" : {
                     "application/json" : {
                        "schema" : {
                           "$ref" : "#/components/schemas/InvalidRequestError"
                        }
                     }
                  },
                  "description" : "The payload is invalid"
               },
               "404" : {
                  "content" : {
                     "application/json" : {
                        "schema" : {
                           "$ref" : "#/components/schemas/DisabledEndpointError"
                        }
                     }
                  },
                  "description" : "The feature is not activated"
               }
            },
            "security" : [],
            "summary" : "Send a one-time password by SMS",
            "tags" : [
               "Authentication"
            ]
         }
      },
      "/signin/passwordless/sms/otp" : {
         "post" : {
            "deprecated" : false,
            "parameters" : [],
            "requestBody" : {
               "content" : {
                  "application/json" : {
                     "schema" : {
                        "$ref" : "#/components/schemas/SignInOtpSchema"
                     }
                  }
               },
               "description" : "",
               "required" : true
            },
            "responses" : {
               "200" : {
                  "content" : {
                     "application/json" : {
                        "schema" : {
                           "$ref" : "#/components/schemas/SessionPayload"
                        }
                     }
                  },
                  "description" : "User successfully authenticated"
               },
               "400" : {
                  "content" : {
                     "application/json" : {
                        "schema" : {
                           "$ref" : "#/components/schemas/InvalidRequestError"
                        }
                     }
                  },
                  "description" : "The payload is invalid"
               },
               "401" : {
                  "content" : {
                     "application/json" : {
                        "schema" : {
                           "$ref" : "#/components/schemas/UnauthorizedError"
                        }
                     }
                  },
                  "description" : "Error processing the request"
               },
               "404" : {
                  "content" : {
                     "application/json" : {
                        "schema" : {
                           "$ref" : "#/components/schemas/DisabledEndpointError"
                        }
                     }
                  },
                  "description" : "The feature is not activated"
               }
            },
            "security" : [],
            "summary" : "Passwordless authentication from a one-time password code received by SMS",
            "tags" : [
               "Authentication"
            ]
         }
      },
...
      "/signin/webauthn" : {
         "post" : {
            "deprecated" : false,
            "parameters" : [],
            "requestBody" : {
               "content" : {
                  "application/json" : {
                     "schema" : {
                        "$ref" : "#/components/schemas/SignInWebauthnSchema"
                     }
                  }
               },
               "description" : "",
               "required" : true
            },
            "responses" : {
               "400" : {
                  "content" : {
                     "application/json" : {
                        "schema" : {
                           "$ref" : "#/components/schemas/InvalidRequestError"
                        }
                     }
                  },
                  "description" : "The payload is invalid"
               },
               "404" : {
                  "content" : {
                     "application/json" : {
                        "schema" : {
                           "$ref" : "#/components/schemas/DisabledEndpointError"
                        }
                     }
                  },
                  "description" : "The feature is not activated"
               }
            },
            "security" : [],
            "summary" : "Sign in using email via FIDO2 Webauthn authentication",
            "tags" : [
               "Authentication"
            ]
         }
      },
      "/signin/webauthn/verify" : {
         "post" : {
            "deprecated" : false,
            "parameters" : [],
            "requestBody" : {
               "content" : {
                  "application/json" : {
                     "schema" : {
                        "$ref" : "#/components/schemas/SignInVerifyWebauthnSchema"
                     }
                  }
               },
               "description" : "",
               "required" : true
            },
            "responses" : {
               "200" : {
                  "content" : {
                     "application/json" : {
                        "schema" : {
                           "$ref" : "#/components/schemas/SessionPayload"
                        }
                     }
                  },
                  "description" : "Signed in successfully"
               },
               "400" : {
                  "content" : {
                     "application/json" : {
                        "schema" : {
                           "$ref" : "#/components/schemas/InvalidRequestError"
                        }
                     }
                  },
                  "description" : "The payload is invalid"
               },
               "401" : {
                  "content" : {
                     "application/json" : {
                        "schema" : {
                           "$ref" : "#/components/schemas/UnauthorizedError"
                        }
                     }
                  },
                  "description" : "Invalid email or password, or user is not verified"
               },
               "404" : {
                  "content" : {
                     "application/json" : {
                        "schema" : {
                           "$ref" : "#/components/schemas/DisabledEndpointError"
                        }
                     }
                  },
                  "description" : "The feature is not activated"
               }
            },
            "security" : [],
            "summary" : "Verfiy FIDO2 Webauthn authentication using public-key cryptography",
            "tags" : [
               "Authentication"
            ]
         }
      },
...
   },
   "security" : [
      {
         "BearerAuth" : []
      }
   ],
   "servers" : [],
...
}